fix(security): address CodeQL high-severity findings in MoRI code
Resolve the high-severity SAST findings flagged by CodeQL on MoRI's own
sources, and make the scan report MoRI-only:

- Integer overflow (CWE-190, x9): widen the first operand to size_t before
  multiplying in size/index computations (dispatch_combine, symmetric_memory,
  shmem/init, ScaleBytes), so the product is computed in 64-bit.
- World-writable files (CWE-732, x6): create barrier/lock/shm files with 0600
  instead of 0666 (same-user inter-process coordination).
- Overly permissive exec perms (CWE-732, x2): install spdk_proxy/umbp_master
  with 0700 instead of 0755.
- Uncontrolled library load (ibv_shim): restrict MORI_IBVERBS_LIB to a bare
  soname (reject path separators) so the override resolves only via the trusted
  linker search path; docs updated to point out-of-tree paths at LD_LIBRARY_PATH.
- Scan scope: post-process SARIF (tools/codeql/filter_sarif.py) to drop
  third-party/test/build results, since CodeQL paths-ignore is not reliably
  applied to compiled C++.

+++ src/application/memory/symmetric_memory.cpp
@@ -240,7 +240,7 @@ SymmMemObjPtr SymmMemManager::RegisterSymmMemObj(void* localPtr, size_t size, bo
       for (size_t q = 0; q < numOfQueuesPerDevice; q++) {
         auto* anvilHandle = anvil::anvil.getSdmaQueue(srcDeviceId, dstDeviceId, q)->deviceHandle();
         HIP_RUNTIME_CHECK(
-            hipMemcpy(&gpuMemObj->deviceHandles_d[dstDeviceId * numOfQueuesPerDevice + q],
+            hipMemcpy(&gpuMemObj->deviceHandles_d[static_cast<size_t>(dstDeviceId) * numOfQueuesPerDevice + q],
                       &anvilHandle, sizeof(anvilHandle), hipMemcpyHostToDevice));
       }
     }
